microservice/pyproject.toml
[tool.poetry]
name = "noise-microsservices"
version = "0.1.0"
description = ""
authors = ["Allan Garcia", "Eduarda Neumann"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
aiokafka = "^0.8.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
fastapi = { extras = ["all"], version = "^0.104.1" }
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
orjson = "^3.9.10"
sqlalchemy = "^2.0.23"
protobuf = "^4.25.1"
psycopg2-binary = "^2.9.9"
aiomqtt = "^1.2.1"
uvloop = "^0.19.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
grpcio-tools = "^1.59.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
